Refactor CurrentState to PebbleStatus and LiveStatus, + fx event sent by LiveService
